etcdserver: move error to errors.go
Both server.go and cluster.go are using defined ErrX. Move error
to errors.go
